commented all classes

+++ app/model/Cart.class.php
@@ -49,9 +49,7 @@ class Cart extends DbObject {
 
     // load all products
     public static function getAllProducts($limit=null) {
-        $query = sprintf(" SELECT id FROM %s ORDER BY id DESC ",
-            self::DB_TABLE
-            );
+        $query = sprintf(" SELECT id FROM %s ORDER BY id DESC ", self::DB_TABLE);
         $db = Db::instance();
         $result = $db->lookup($query);
         if(!mysql_num_rows($result))
@@ -65,31 +63,25 @@ class Cart extends DbObject {
         }
     }
 
-    //remove product issues
+    //remove products from cart
     public function removeProduct($id, $uid){
-      $query = sprintf(" DELETE FROM %s
-                          WHERE id=$id AND creator_id=$uid; ",
-          (string)self::DB_TABLE
-          );
+      $query = sprintf(" DELETE FROM %s WHERE id=$id AND creator_id=$uid; ",
+          (string)self::DB_TABLE);
       $db = Db::instance();
       $db->execute($query);
     }
 
-    //add product issues
+    //add product to cart
     public function addProduct($title, $price, $img, $uid){
-      /**$query = sprintf("INSERT INTO %s (`title`, `description`, `price`, `sizes`, `image_url`)
-                         VALUES (%s, %s, %s, %s, %s);",
-                         (string)self::DB_TABLE, $title, $desc, $sizes, $price, $img
-                       );**/
       $query = "INSERT INTO cart VALUES (DEFAULT, '$uid', '$title', '$img', '$price')";
       $db = Db::instance();
       $db->execute($query);
     }
 
+    //get the items by the user who added them
     public function getProductsByUser($id) {
       $query = sprintf(" SELECT id FROM %s WHERE creator_id={$id} ORDER BY id DESC ",
-          self::DB_TABLE
-          );
+          self::DB_TABLE);
       $db = Db::instance();
       $result = $db->lookup($query);
       if(!mysql_num_rows($result))
@@ -102,5 +94,4 @@ class Cart extends DbObject {
           return ($objects);
       }
     }
-
 }
